:truck: chore: cleaning up shared module
